test: update TarDir unit test
The current TestTarDir does not check if an empty archive is produced. Update
the test case to check for empty tar archives.

The include test case does not correctly check for *.mod, files. Update the
assertion to expect files matching either pattern.

The test is currently failing and will be fixed in the next few commits.

Signed-off-by: Aaron D Borden <adborden@a14n.net>